RTEMIS-180: PR modifications as per the PR review.
